requirements.txt
flask[async]
sqlalchemy
hypercorn
python-dotenv